Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,24 +1,24 @@
 #include <stdio.h>
 int main() {
-    int n1,n2;
+    float n1,n2;
     char ch;
-    scanf("%d %d %c",&n1,&n2,&ch);
+    scanf("%f %f %c",&n1,&n2,&ch);
     switch (ch){
         case '+':
-            printf("%d",n1+n2);
+            printf("%.0f",n1+n2);
             break;
         case '-':
             if (n1>n2){
-                printf("%d",n1-n2);
+                printf("%.0f",n1-n2);
                 break;
             }
             else printf("error");
         case '*':
-            printf("%d",n1*n2);
+            printf("%.0f",n1*n2);
             break;
         case '/':
             if (n1>n2){
-                printf("%d",n1/n2);
+                printf("%.0f",n1/n2);
                 break;
             }
             else printf("erorr");
